Revert "Move db.js to a common directory, as migrate will use it as well"
This reverts commit 2b14725c6867ede52574406257c75b9be0dd67b1.
Because the migration system will be a scripting system while the normal
database usage is in a server system, the two do not have enough overlap
to be very worthwhile in treating as common code.
